Refactor DFE Sign In related code
This work is part of an overhaul of how we structure/namespace the code
related with external integrations prior to add new integrations to the
service.

This is a refactor, so the behaviour should stay the same.
The principles behind the refactor are:
- Define clear boundaries between usage of "lib" and "app/services"
  project folders:
  - The library folder is left to domain agnostic code that could be
    extracted into a gem and doesn't depend on our service domain.
    In this case the library folder is exclusively used for  DFE Sign In
    API related code.
  - The services folder is used for code that depends on our domain.
    In this case we use for code relating to parsing, exporting and
    writing into DB the DSI users info obtained from the API calls.

- Structure the DSI related code following Object Oriented Programming
  principles.
  - Instead of having a collection of methods around, we grouped and
    extracted different DSI related objects with different
    responsibilities.
  - Used modules/classes to namespace and group the DSI code.
