+++ app/jobs/update_dsi_users_in_db_job.rb
@@ -1,9 +1,7 @@
-require "update_dsi_users_in_db"
-
 class UpdateDSIUsersInDbJob < ApplicationJob
   queue_as :low
 
   def perform
-    UpdateDSIUsersInDb.new.run!
+    Publishers::DfeSignIn::UpdateUsersInDb.new.call
   end
 end
