+++ config/environments/test.rb
@@ -64,9 +64,9 @@ Rails.application.configure do
   # Use test geocoder lookup, unless otherwise specified
   config.geocoder_lookup = :test
 
-  require "fake_dsi_sign_out_endpoint"
+  require "dfe_sign_in/fake_sign_out_endpoint"
   ENV["DFE_SIGN_IN_ISSUER"] = "http://fake.dsi.example.com"
-  config.middleware.insert_before 0, FakeDSISignOutEndpoint
+  config.middleware.insert_before 0, DfeSignIn::FakeSignOutEndpoint
 
   config.log_file_size = 100.megabytes
 end
