+++ spec/jobs/export_dsi_users_to_big_query_job_spec.rb
@@ -9,13 +9,13 @@ RSpec.describe ExportDSIUsersToBigQueryJob do
     let(:disable_expensive_jobs_enabled?) { false }
 
     it "invokes the libs to export users and approvers to big query" do
-      export_dsi_users_to_big_query = double(:export_dsi_users_to_big_query)
-      expect(ExportDSIUsersToBigQuery).to receive(:new) { export_dsi_users_to_big_query }
-      expect(export_dsi_users_to_big_query).to receive(:run!)
+      export_dsi_users_to_big_query = instance_double(Publishers::DfeSignIn::BigQueryExport::Users)
+      expect(Publishers::DfeSignIn::BigQueryExport::Users).to receive(:new) { export_dsi_users_to_big_query }
+      expect(export_dsi_users_to_big_query).to receive(:call)
 
-      export_dsi_approvers_to_big_query = double(:export_dsi_approvers_to_big_query)
-      expect(ExportDSIApproversToBigQuery).to receive(:new) { export_dsi_approvers_to_big_query }
-      expect(export_dsi_approvers_to_big_query).to receive(:run!)
+      export_dsi_approvers_to_big_query = instance_double(Publishers::DfeSignIn::BigQueryExport::Approvers)
+      expect(Publishers::DfeSignIn::BigQueryExport::Approvers).to receive(:new) { export_dsi_approvers_to_big_query }
+      expect(export_dsi_approvers_to_big_query).to receive(:call)
 
       perform_enqueued_jobs { job }
     end
@@ -25,8 +25,8 @@ RSpec.describe ExportDSIUsersToBigQueryJob do
     let(:disable_expensive_jobs_enabled?) { true }
 
     it "does not perform the job" do
-      expect(ExportDSIUsersToBigQuery).not_to receive(:new)
-      expect(ExportDSIApproversToBigQuery).not_to receive(:new)
+      expect(Publishers::DfeSignIn::BigQueryExport::Users).not_to receive(:new)
+      expect(Publishers::DfeSignIn::BigQueryExport::Approvers).not_to receive(:new)
 
       perform_enqueued_jobs { job }
     end
